bug fixes, slight lighting changes
Fixed a bug that caused the spoiler upgrade to not appear.
Fixed a bug that caused the hood upgrade to not appear.
Fixed a bug that caused the player to be set backwards at a certain checkpoint.
Fixed a bug that caused only the acceleration tier number to appear in the shops.
Made the lightning so that it isn't so dark you cannot see anything in the turns